Move all nightly tests to test/registered/ and delete test/nightly/
- Move Ascend/NPU tests to test/registered/ascend/
  - llm_models/ (17 tests + gsm8k_ascend_mixin.py)
  - vlm_models/ (10 tests + mmmu-val.yaml)
  - embedding_models/ (1 test)
  - rerank_models/ (1 test)
- Move utility modules to python/sglang/test/:
  - accuracy_test_runner.py
  - performance_test_runner.py
  - run_combined_tests.py
- Update imports in 8-gpu-models tests to use sglang.test.*
- Update mmmu-val.yaml path in test_vlm_utils.py
- Delete empty test/nightly/ directory
