+++ test/registered/ascend/embedding_models/test_ascend_embedding_models.py
