+++ test/registered/ascend/llm_models/test_ascend_afm_4_5b.py
