+++ test/registered/ascend/llm_models/test_ascend_baichuan2_13b_chat.py
