+++ test/registered/ascend/llm_models/test_ascend_c4ai_command_r_v01.py
