+++ test/registered/ascend/llm_models/test_ascend_charglm2_6b.py
