+++ test/registered/ascend/llm_models/test_ascend_exaone_3.py
