+++ test/registered/ascend/llm_models/test_ascend_glm4_9b_chat.py
