+++ test/registered/ascend/llm_models/test_ascend_granite_3_0_3b_a800m.py
