+++ test/registered/ascend/llm_models/test_ascend_granite_3_1_8b.py
