+++ test/registered/ascend/llm_models/test_ascend_internlm2_7b.py
