+++ test/registered/ascend/llm_models/test_ascend_llama_2_7b.py
