+++ test/registered/ascend/llm_models/test_ascend_mistral_7b.py
