+++ test/registered/ascend/llm_models/test_ascend_persimmon_8b_chat.py
