+++ test/registered/ascend/llm_models/test_ascend_phi_4_multimodal.py
