+++ test/registered/ascend/llm_models/test_ascend_smollm_1_7b.py
