+++ test/registered/ascend/llm_models/tool_chat_template_c4ai_command_r_v01.jinja
