+++ test/registered/ascend/rerank_models/test_ascend_cross_encoder_models.py
