+++ test/registered/ascend/vlm_models/mmmu-val.yaml
