+++ test/registered/ascend/vlm_models/test_ascend_gemma_3_4b_it.py
