+++ test/registered/ascend/vlm_models/test_ascend_janus_pro_1b.py
