+++ test/registered/ascend/vlm_models/test_ascend_janus_pro_7b.py
