+++ test/registered/ascend/vlm_models/test_ascend_minicpm_o_2_6.py
