+++ test/registered/ascend/vlm_models/test_ascend_minicpm_v_2_6.py
