+++ test/registered/ascend/vlm_models/test_ascend_phi4_multimodal_instruct.py
