+++ test/registered/ascend/vlm_models/test_ascend_qwen2_5_vl_3b_instruct.py
